requirements.txt
Flask_SQLAlchemy==2.4.1
lxml==4.4.1
matplotlib==3.1.1
PyMySQL==0.9.3
wordcloud==1.6.0
SQLAlchemy==1.3.9
Flask==1.1.1
tqdm==4.36.1
requests==2.22.0
numpy==1.16.5
apscheduler==3.6.3
beautifulsoup4==4.9.1
flask_wtf==0.14.3
jieba==0.42.1
Pillow==7.1.2
snownlp==0.12.3
WTForms==2.3.1
jinja2==3.0.3
werkzeug==2.0.2
itsdangerous==2.0.1
mysql-connector-python